started moving create note to class
