+++ src-ui/renderer/reactive/sidebar/index.ts
@@ -1,2 +1,1 @@
-export { renderSidebarCreateNote } from './sidebar-create-note'
 export { sidebar } from './sidebar'
